Introducing a flag to fix MAF revcomp issue while link import
This flag is used while import of all links to determine if MAF was
the input file. In that case, we need to adjust the location of hits
on (-)strand. This seems to be a bug in BioPerl and might be fixed in
future.

An issue for BioPerl was already opened:

	https://github.com/bioperl/bioperl-live/issues/147